+++ src/features/manage/AccountSettingsCard/AccountSettingsCard.tsx
@@ -1,4 +1,4 @@
-import { GetProfileResponse } from '@/models/manage.types';
+import { GetProfileResponse } from '@/features/manage/manage.types';
 import { Card, CardBody, CardTitle } from '@/components/Card/Card';
 import { ButtonGroup } from '@/components/ButtonGroup/ButtonGroup';
 import { CardSkeleton } from '@/components/Skeleton/Skeleton';
