+++ src/pages/change-password/index.tsx
@@ -1,6 +1,9 @@
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
-import { useChangePasswordMutation, useGetProfileQuery } from '@/redux/api';
+import {
+    useGetProfileQuery,
+    useChangePasswordMutation
+} from '@/features/manage/manage.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -13,8 +16,9 @@ import {
 
 import { GetServerSideProps } from 'next';
 import { getServerSession } from 'next-auth';
-import { getProfile, getRunningQueriesThunk } from '@/redux/api';
+import { getRunningQueriesThunk } from '@/redux/api';
 import { wrapper } from '@/redux/store';
+import { getProfile } from '@/features/manage/manage.api';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
 
 const ChangePasswordPage = () => {
