+++ src/pages/forgot-password/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
-import { useForgotPasswordMutation } from '@/redux/api';
+import { useForgotPasswordMutation } from '@/features/account/account.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -20,7 +20,7 @@ const ForgotPasswordPage = () => {
         toast.success(
             'Instructions as to how to reset your password have been sent to you via email.'
         );
-    
+
         await router.push('/login');
     };
