+++ src/pages/my-account/index.tsx
@@ -1,6 +1,9 @@
 import { signOut } from 'next-auth/react';
 import toast from 'react-hot-toast';
-import { useDeleteAccountMutation, useGetProfileQuery } from '@/redux/api';
+import {
+    useGetProfileQuery,
+    useDeleteAccountMutation
+} from '@/features/manage/manage.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -10,7 +13,8 @@ import { AccountSettingsCard } from '@/features/manage/AccountSettingsCard/Accou
 
 import { GetServerSideProps } from 'next';
 import { getServerSession } from 'next-auth';
-import { getRunningQueriesThunk, getProfile } from '@/redux/api';
+import { getRunningQueriesThunk } from '@/redux/api';
+import { getProfile } from '@/features/manage/manage.api';
 import { wrapper } from '@/redux/store';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
