+++ src/pages/register/index.tsx
@@ -1,6 +1,6 @@
 import { signIn } from 'next-auth/react';
 import toast from 'react-hot-toast';
-import { useRegisterMutation } from '@/redux/api';
+import { useRegisterMutation } from '@/features/account/account.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -15,7 +15,7 @@ const RegisterPage = () => {
 
     const onSubmit = async (values: RegisterFormValues) => {
         await register(values).unwrap();
-        toast.success('User successfully registered.');  
+        toast.success('User successfully registered.');
         await signIn('credentials', { ...values, callbackUrl: '/' });
     };
