+++ src/pages/reset-password/[token].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import toast from 'react-hot-toast';
-import { useResetPasswordMutation } from '@/redux/api';
+import { useResetPasswordMutation } from '@/features/account/account.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
