+++ src/pages/user/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import { UserSort } from '@/models/user.types';
-import { useSearchUsersQuery } from '@/redux/api';
+import { UserSort } from '@/features/user/user.types';
+import { useSearchUsersQuery } from '@/features/user/user.api';
 import { Meta } from '@/components/Meta/Meta';
 import { Container } from '@/components/Container/Container';
 import { Title } from '@/components/Title/Title';
@@ -16,7 +16,8 @@ import { UserList } from '@/features/user/UserList/UserList';
 
 import { GetServerSideProps } from 'next';
 import { getServerSession } from 'next-auth';
-import { getRunningQueriesThunk, searchUsers } from '@/redux/api';
+import { getRunningQueriesThunk } from '@/redux/api';
+import { searchUsers } from '@/features/user/user.api';
 import { wrapper } from '@/redux/store';
 import { authOptions } from '@/pages/api/auth/[...nextauth]';
