+++ src/utils/handler.ts
@@ -1,7 +1,6 @@
 import { NextApiHandler, NextApiRequest, NextApiResponse } from 'next';
 import { getToken } from 'next-auth/jwt';
 import { ValidationError } from 'yup';
-import { ProblemResponse } from '@/models/base.types';
 import { Role, isAuthorized } from '@/utils/auth';
 import { config } from '@/utils/config';
 
@@ -11,7 +10,7 @@ type HandlerContext = {
 
 export type Handler<TResponse = unknown> = (
     req: NextApiRequest,
-    res: NextApiResponse<TResponse | ProblemResponse>,
+    res: NextApiResponse<TResponse>,
     context: HandlerContext
 ) => unknown | Promise<unknown>;
